+++ app/pages/hungary/index.php
@@ -27,7 +27,6 @@
             <?= \Hungary\ElectionResultContainer::show(
                 election: "2026",
                 title: ["2026", "Parliamentary", "Election"],
-                live: TRUE
             ); ?>
             <?= \Hungary\ElectionResultContainer::show(
                 election: "2022",