+++ app/pages/hungary/parliamentary-elections/index.php
@@ -25,8 +25,7 @@
             <?= \Hungary\ElectionResultContainer::show(
                 election: "2026",
                 title: ["2026", "Parliamentary", "Election"],
-                messages: ['group' => "2026", 'open' => TRUE],
-                live: TRUE
+                messages: ['group' => "2026", 'open' => TRUE]
             ); ?>
 
             <?= \Hungary\ElectionResultContainer::show(