+++ app/ssr-components/canada/ElectionResultContainer/ElectionResultContainer.php
@@ -2,13 +2,15 @@
 namespace Canada;
 
 class ElectionResultContainer extends \Shared\ElectionResultContainer{
+    static $country = "canada";
+    
     static function render (
         string $election,
         array $title,                               // [string, string, string]
         ?array $dataAttrs = [],                     // [{name} => value: any]
         ?array $messages = [],                      // [group: string, open: bool?]
         ?bool $showChanges = FALSE,
-        ?bool $live = FALSE,
+        ?bool $live = NULL,
         ?string $dedicatedPage = NULL,
         ?string $winFormulaName = "default",
         ?string $regionsType = NULL