+++ app/ssr-components/uk/RegionPage/ScotlandRegionPage.php
@@ -46,12 +46,12 @@ class Scotland extends \UK\RegionPage{
                 case "S2011": // full list candidate data is not available for S2011 ?>
                     <?= \Shared\ElectedCandidatesMasonryList::render($results); ?>
                     <h3>Votes</h3>
-                    <?= \Shared\DHondtTable::render($results, $divisors ?? []); ?>
+                    <?= \Shared\DHondtTable::render($results, $divisors ?? [], fixedRounds: 7); ?>
                     <?php break; ?>
                 <?php default: ?>
                     <?= \Shared\CandidatesMasonryList::render($results); ?>
                     <h3>Votes</h3>
-                    <?= \Shared\DHondtTable::render($results, $divisors ?? []); ?>
+                    <?= \Shared\DHondtTable::render($results, $divisors ?? [], fixedRounds: 7); ?>
             <?php endswitch; ?>
         </article>