+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 5.2 - 2026-04-24
+- Add 2026 Scottish Parliament and Senedd Cymru elections
+- Start and end live events from database
+- Delay drawing poll graph lines until at least three recent polls exist
+- Change page title separator from pipe to hyphen
+
 ## 5.1 - 2026-04-12
 - Add support for live events